Fix erroring tests
